services/financial-engine/pyproject.toml
[build-system]
requires = ["setuptools>=72", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cartool-financial-engine"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "pip-audit>=2.7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
